Update ImageResize.php
This is a solution for Yii2 advanced.

The developer must write two cache paths in the configuration: ['assets/images', '../../frontend/web/assets/images']

The image resizer will create two caches, one for assets in the backend and another for the assets in the frontend., will prioritize the first provided path that, by default, will be the right path in both.

So TinyMCE will show the image in the backend and the image can be showed in frontend as well. I also move the quality from 50 to 80. Nobody will complain about that in this times. :)

Solution for:  Not working in Yii2 Advance template #35  - https://github.com/noam148/yii2-image-manager/issues/35